# api_docs.yaml
...
info:
  description: powered by Flasgger
  termsOfService: "/tos"
  title: A swagger API
  version: 0.0.1
paths:
...
  "/admin/users/manage":
    post:
      description: Manage User
      parameters:
        - description: User ID
          in: query
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            text/html:
              example: User managed successfully
          description: User managed successfully
        "400":
          content:
            text/html:
              example: User not found
          description: User not found
      summary: Manage User
      tags:
        - Admin
  "/admin/users/view":
    get:
      description: View User
      parameters:
        - description: User ID
          in: query
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            text/html:
              example: View User
          description: View User
        "400":
          content:
            text/html:
              example: User not found
          description: User not found
      summary: View User
      tags:
        - Admin
...
swagger: "2.0"